Discord_Bot/requirements.txt
discord.py
python-dotenvp